[FLINK-21942] Extract JobLeaderIdService interface to make the ResourceManager better testable
